backend/requirements.txt
Flask==3.0.0
celery==5.3.4
redis==5.0.1
PyMuPDF==1.24.0
EbookLib==0.18.0
pytesseract==0.3.10
Pillow==10.0.0
langdetect==1.0.9
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
psycopg2-binary==2.9.9
PyJWT==2.8.0
Flask-Limiter==3.5.0
prometheus-client==0.20.0
python-magic==0.4.27
pypandoc==1.15
pdf2image==1.16.3
camelot-py==0.11.0
supabase==2.0.0

pytest==7.4.3
pytest-flask==1.3.0
